+++ spec/default_spec.rb
@@ -1,19 +1,32 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe 'yum-nginx' do
-  describe 'when on redhat 7.0' do
+  describe 'when on redhat 7.4' do
     describe 'by default' do
       let(:chef_run) do
-        ChefSpec::SoloRunner.new(platform: 'redhat', version: '7.0')
-          .converge(described_recipe)
+        ChefSpec::SoloRunner.new.converge(described_recipe)
+      end
+
+      it 'downloaded gpg key file and notify resource to import key' do
+        resource = chef_run.cookbook_file('nginx_signing.key')
+        expect(chef_run).to create_cookbook_file('nginx_signing.key')
+          .with(path: "#{Dir.tmpdir}/nginx_signing.key")
+        expect(resource).to notify('execute[import gpgkey]').to(:run)
+                                                            .immediately
+      end
+
+      it 'imported gpg key' do
+        expect(chef_run.execute('import gpgkey')).to do_nothing
       end
 
       it 'should create the nginx-stable repo with default attribs' do
         expect(chef_run).to create_yum_repository('nginx-stable').with(
           description: 'nginx stable repo',
-          baseurl: 'http://nginx.org/packages/redhat/7/$basearch/',
-          gpgcheck: false,
-          sslverify: false,
+          baseurl: 'https://nginx.org/packages/redhat/7/$basearch/',
+          gpgcheck: true,
+          sslverify: true,
           enabled: true
         )
       end
@@ -33,11 +46,11 @@ describe 'yum-nginx' do
 
     describe 'when nginx-stable is not managed and non-stable repos managed' do
       let(:chef_run) do
-        ChefSpec::SoloRunner.new(platform: 'redhat', version: '7.0') do |node|
-          node.set['yum-nginx']['repos']['nginx-stable']['managed'] = false
-          node.set['yum-nginx']['repos']['nginx-stable-source']['managed'] = true
-          node.set['yum-nginx']['repos']['nginx-mainline']['managed'] = true
-          node.set['yum-nginx']['repos']['nginx-mainline-source']['managed'] = true
+        ChefSpec::SoloRunner.new do |node|
+          node.normal['yum-nginx']['repos']['nginx-stable']['managed'] = false
+          node.normal['yum-nginx']['repos']['nginx-stable-source']['managed'] = true
+          node.normal['yum-nginx']['repos']['nginx-mainline']['managed'] = true
+          node.normal['yum-nginx']['repos']['nginx-mainline-source']['managed'] = true
         end.converge(described_recipe)
       end
 
@@ -48,9 +61,9 @@ describe 'yum-nginx' do
       it 'should create the nginx-stable-source repo with default attribs' do
         expect(chef_run).to create_yum_repository('nginx-stable-source').with(
           description: 'nginx stable source repo',
-          baseurl: 'http://nginx.org/packages/redhat/7/SRPMS/',
-          gpgcheck: false,
-          sslverify: false,
+          baseurl: 'https://nginx.org/packages/redhat/7/SRPMS/',
+          gpgcheck: true,
+          sslverify: true,
           enabled: true
         )
       end
@@ -58,9 +71,9 @@ describe 'yum-nginx' do
       it 'should create the nginx-mainline repo with default attribs' do
         expect(chef_run).to create_yum_repository('nginx-mainline').with(
           description: 'nginx mainline repo',
-          baseurl: 'http://nginx.org/packages/mainline/redhat/7/$basearch/',
-          gpgcheck: false,
-          sslverify: false,
+          baseurl: 'https://nginx.org/packages/mainline/redhat/7/$basearch/',
+          gpgcheck: true,
+          sslverify: true,
           enabled: true
         )
       end
@@ -68,9 +81,9 @@ describe 'yum-nginx' do
       it 'should create the nginx-mainline-source repo default attribs' do
         expect(chef_run).to create_yum_repository('nginx-mainline-source').with(
           description: 'nginx mainline source repo',
-          baseurl: 'http://nginx.org/packages/mainline/redhat/7/SRPMS/',
-          gpgcheck: false,
-          sslverify: false,
+          baseurl: 'https://nginx.org/packages/mainline/redhat/7/SRPMS/',
+          gpgcheck: true,
+          sslverify: true,
           enabled: true
         )
       end
@@ -78,15 +91,15 @@ describe 'yum-nginx' do
 
     describe 'when adding a custom nginx repo' do
       let(:chef_run) do
-        ChefSpec::SoloRunner.new(platform: 'redhat', version: '7.0') do |node|
-          node.set['yum-nginx']['repos']['nginx-new']['managed'] = true
-          node.set['yum-nginx']['repos']['nginx-new']['repositoryid'] = 'nginx-new'
-          node.set['yum-nginx']['repos']['nginx-new']['description'] = 'nginx New'
-          node.set['yum-nginx']['repos']['nginx-new']['baseurl'] = 'http://path.to.nginx'
-          node.set['yum-nginx']['repos']['nginx-new']['gpgcheck'] = true
-          node.set['yum-nginx']['repos']['nginx-new']['gpgkey'] = 'file://path/to/gpg/key'
-          node.set['yum-nginx']['repos']['nginx-new']['sslverify'] = false
-          node.set['yum-nginx']['repos']['nginx-new']['enabled'] = true
+        ChefSpec::SoloRunner.new do |node|
+          node.normal['yum-nginx']['repos']['nginx-new']['managed'] = true
+          node.normal['yum-nginx']['repos']['nginx-new']['repositoryid'] = 'nginx-new'
+          node.normal['yum-nginx']['repos']['nginx-new']['description'] = 'nginx New'
+          node.normal['yum-nginx']['repos']['nginx-new']['baseurl'] = 'http://path.to.nginx'
+          node.normal['yum-nginx']['repos']['nginx-new']['gpgcheck'] = true
+          node.normal['yum-nginx']['repos']['nginx-new']['gpgkey'] = 'file://path/to/gpg/key'
+          node.normal['yum-nginx']['repos']['nginx-new']['sslverify'] = true
+          node.normal['yum-nginx']['repos']['nginx-new']['enabled'] = true
         end.converge(described_recipe)
       end
 
@@ -97,18 +110,17 @@ describe 'yum-nginx' do
           baseurl: 'http://path.to.nginx',
           gpgcheck: true,
           gpgkey: 'file://path/to/gpg/key',
-          sslverify: false,
+          sslverify: true,
           enabled: true
         )
       end
     end
   end
 
-  ['5.10', '6.0', '7.0'].each do |version|
+  ['6.9', '7.4'].each do |version|
     describe "when on redhat #{version}" do
       let(:chef_run) do
-        ChefSpec::SoloRunner.new(platform: 'redhat', version: version)
-          .converge(described_recipe)
+        ChefSpec::SoloRunner.new(platform: 'redhat', version: version).converge(described_recipe)
       end
 
       it 'should not raise an error' do
@@ -119,13 +131,13 @@ describe 'yum-nginx' do
 
   describe 'when on an unsupported platform' do
     let(:chef_run) do
-      ChefSpec::SoloRunner.new(platform: 'redhat', version: '7.0') do |node|
-        node.set['yum-nginx']['rhel']['supported-versions']['7'] = false
+      ChefSpec::SoloRunner.new do |node|
+        node.normal['yum-nginx']['rhel']['supported-versions']['7'] = false
       end.converge(described_recipe)
     end
 
     it 'should raise an error' do
-      expect { chef_run }.to raise_error('rhel/redhat/7.0 is not supported by the default recipe')
+      expect { chef_run }.to raise_error('rhel/redhat/7.4 is not supported by the default recipe')
     end
   end
 end