+++ test/fixtures/cookbooks/test/recipes/install_nginx_source.rb
@@ -1,1 +1,3 @@
+# frozen_string_literal: true
+
 execute 'yumdownloader --source nginx'