+++ test/fixtures/cookbooks/test/recipes/yum_utils.rb
@@ -1,1 +1,3 @@
+# frozen_string_literal: true
+
 package 'yum-utils' if platform_family?('rhel')