+++ test/integration/default/serverspec/default_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 # nginx package should be installed when a repository is enabled