+++ test/integration/mainline_source/serverspec/mainline_source_spec.rb
@@ -1,7 +1,9 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 # the assumption made here is that if there is a source repo present, then the
 #  the source package was downloaded successfully
-describe command('ls /nginx-*.ngx.src.rpm') do
+describe command('yum repolist -v nginx-mainline-source') do
   its(:exit_status) { should eq 0 }
 end